requirements.txt
altgraph==0.17.4
annotated-types==0.6.0
colorama==0.4.6
contourpy==1.2.0
coverage==7.4.3
customtkinter==5.2.1
cycler==0.12.1
darkdetect==0.8.0
et-xmlfile==1.1.0
fonttools==4.49.0
iniconfig==2.0.0
kiwisolver==1.4.5
matplotlib==3.8.3
mpmath==1.3.0
multimethod==1.11.2
mypy==1.8.0
mypy-extensions==1.0.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.2.1
pandas-stubs==2.2.0.240218
pandera==0.18.0
pefile==2023.2.7
Pillow==10.1.0
pluggy==1.4.0
psutil==5.9.8
pydantic==2.6.3
pydantic_core==2.16.3
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2024.0
pyparsing==3.1.2
pytest==8.0.2
python-dateutil==2.8.2
pytz==2023.3.post1
pywin32-ctypes==0.2.2
scipy==1.11.4
setuptools==69.0.3
six==1.16.0
sympy==1.12
tk==0.1.0
tkinter-tooltip==2.2.0
typeguard==4.1.5
types-Pillow==10.2.0.20240311
types-pytz==2024.1.0.20240203
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
wrapt==1.16.0
